move `generateMock` into `MockState.ts`
